moved enums out of dataconverter and into options package
