+++ src/main/java/com/usf/utils/options/CardOption.java
@@ -1,0 +1,6 @@
+package com.usf.utils.options;
+
+public enum CardOption {
+    FORMATTED, COMMON, COMMON_FORMATTED,
+    FULL_YEAR, EXPIRED
+}
