+++ src/main/java/com/usf/utils/options/DateOption.java
@@ -1,0 +1,5 @@
+package com.usf.utils.options;
+
+public enum DateOption {
+    FUTURE, PAST
+}
