+++ src/main/java/com/usf/utils/options/NameOption.java
@@ -1,0 +1,7 @@
+package com.usf.utils.options;
+
+public enum NameOption {
+    FIRSTNAME_ONLY, LASTNAME_ONLY,
+    WITH_MIDDLENAME, WITH_TITLE,
+    LASTNAME_FIRST
+}
