+++ src/main/java/com/usf/utils/options/PersonOption.java
@@ -1,0 +1,5 @@
+package com.usf.utils.options;
+
+public enum PersonOption {
+    CHILD, TEEN, ADULT, SENIOR
+}
