+++ src/main/java/com/usf/utils/options/StringOption.java
@@ -1,0 +1,6 @@
+package com.usf.utils.options;
+
+public enum StringOption {
+    SYMBOL, ALPHA, DIGIT,
+    WORD, SENTENCE, PARAGRAPH
+}
