+++ src/main/java/com/usf/utils/options/TimeOption.java
@@ -1,0 +1,5 @@
+package com.usf.utils.options;
+
+public enum TimeOption {
+    DAYS, WEEKS, MONTHS, YEARS
+}
